Saving updates

+++ humble-video-captive/src/main/gnu/librtmp/csrc/src/librtmp/rtmp.c
@@ -189,7 +189,7 @@ int
 RTMPPacket_Alloc(RTMPPacket *p, uint32_t nSize)
 {
   char *ptr;
-  if (nSize > SIZE_MAX - RTMP_MAX_HEADER_SIZE)
+  if ((uint64_t)nSize > SIZE_MAX - RTMP_MAX_HEADER_SIZE)
     return FALSE;
   ptr = calloc(1, nSize + RTMP_MAX_HEADER_SIZE);
   if (!ptr)
